+++ packages/form/src/SelectRow.js
@@ -28,6 +28,7 @@ const ROOT_BEM = icBEM(COMPONENT_NAME);
 export const BEM = {
     root: ROOT_BEM,
     popover: ROOT_BEM.element('popover'),
+    placeholder: ROOT_BEM.element('placeholder'),
 };
 
 export const Popover = renderToLayer(
@@ -64,6 +65,7 @@ class SelectRow extends PureComponent {
         asideSeparator: PropTypes.string,
         disabled: PropTypes.bool,
         // <SelectList> props
+        multiple: SelectList.propTypes.multiple,
         values: SelectList.propTypes.values,
         defaultValues: SelectList.propTypes.defaultValues,
         onChange: PropTypes.func,
@@ -78,6 +80,7 @@ class SelectRow extends PureComponent {
         asideSeparator: ', ',
         disabled: false,
         // <SelectList> props
+        multiple: SelectList.defaultProps.multiple,
         values: SelectList.defaultProps.values,
         defaultValues: SelectList.defaultProps.defaultValues,
         onChange: () => {},
@@ -136,16 +139,18 @@ class SelectRow extends PureComponent {
     }
 
     renderRowValuesAside() {
-        const { asideAll, asideNone, asideSeparator } = this.props;
+        const { multiple, asideAll, asideNone, asideSeparator } = this.props;
         const { cachedValues, valueLabelMap } = this.state;
 
-        // Can turn off 'All' display by passing `null`.
-        if (asideAll && cachedValues.length === valueLabelMap.size) {
-            return asideAll;
+        if (cachedValues.length === 0) {
+            return <span className={BEM.placeholder.toString()}>{asideNone}</span>;
         }
 
-        if (cachedValues.length === 0) {
-            return asideNone;
+        if (multiple) {
+            // Can turn off 'All' display by passing `null`.
+            if (asideAll && cachedValues.length === valueLabelMap.size) {
+                return asideAll;
+            }
         }
 
         return cachedValues
@@ -156,8 +161,12 @@ class SelectRow extends PureComponent {
     render() {
         const {
             label,
+            asideAll,
+            asideNone,
+            asideSeparator,
             disabled,
             // <ListRow> props (intercepted from it)
+            // multiple,
             values,
             defaultValues,
             onChange,