+++ packages/form/src/__tests__/TextInputRow.test.js
@@ -51,4 +51,23 @@ describe('Pure <TextInputRow>', () => {
         expect(wrapper.state('focused')).toBeFalsy();
         expect(wrapper.hasClass(focusedModifier)).toBeFalsy();
     });
+
+    it('accepts additional children', () => {
+        const wrapper = mount(
+            <PureTextInputRow label="foo">
+                <span data-foo />
+            </PureTextInputRow>
+        );
+
+        expect(wrapper.containsMatchingElement(<span data-foo />)).toBeTruthy();
+    });
+
+    it('keeps a ref to the <input> inside', () => {
+        const wrapper = mount(
+            <PureTextInputRow label="foo" />
+        );
+
+        const inputRef = wrapper.instance().getInputNode();
+        expect(inputRef).toBeInstanceOf(HTMLInputElement);
+    });
 });