+++ packages/storybook/examples/Icon/InlineIcons.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import Icon from '@ichef/gypcrete/src/Icon';
+
+function InlineIconsSet() {
+    return (
+        <div>
+            <Icon type="inline-loading" />
+            <Icon type="inline-success" />
+            <Icon type="inline-error" />
+            <Icon type="inline-question" />
+        </div>
+    );
+}
+
+export default InlineIconsSet;